journal: shrink EntryItem to 32-bits
In a journal file, data objects such as
"_BOOT_ID=100b7ef4c5124bc6839e28db763b4627" and "PRIORITY=6" are
referenced repeatedly; usually the only difference between two entries
is the "MESSAGE=" data object.  This results in a large amount of wasted
space, as each entry contains an array of 16-byte EntryItem pointers to
such data objects.

A journal file is typically limited to 128MB in size, so using a 64-bit
offset to refer to them is wasting the top 32-bits storing 0x00000000.