sd-journal: implement most operations atop RocksDB
This represents the objects that would be in the journal file natively
as key-value pairs in RocksDB, with a few different types of keys,
prefixed by a type (where || is the concatenation operator):

  * INDEX_ENTRY_BY_OFFSET || <64-bit big-endian integer>
    key represents the "offset" of the entry - which would be the
    identifier that a reader would use to find a specific entry
    value is that entry, the same bytes that would be an EntryObject in
    the mmap'ed binary file format.

  * INDEX_DATA_BY_OFFSET || <64-bit big endian integer>
    key is the "offset" of the data object in the same fashion.
    value is the payload of the data object.  The object header is
    synthesized at read-time.

  * SET_FIELD || <field name>
    key is the name of a field
    value is empty
    This supports `journalctl --fields`.

  * DATA_BY_CONTENTS || <KEY=VALUE data blob>
    key is the KEY=VALUE that was inserted into the database
    value is the offset of the corresponding data object -- i.e. the
    integer in the corresponding INDEX_DATA_BY_OFFSET row.

  * SET_ENTRY_BY_DATA      || <64-bit big endian> || <64-bit big endian>
  * SET_ENTRY_BY_SEQNUM    || <64-bit big endian> || <64-bit big endian>
  * SET_ENTRY_BY_REALTIME  || <64-bit big endian> || <64-bit big endian>
  * SET_ENTRY_BY_MONOTONIC || <64-bit big endian> || <64-bit big endian>
    key is the {"offset" of a data object, sequence number, realtime
    timestamp, or monotonic timestamp} respectively, followed by the
    "offset" of an entry that contains it.
    value is empty; the existence of the row is all that's needed to
    indicate "membership" in the particular set

    This is used to make finding entries that contain a particular field
    value (e.g. `journalctl -u foo.service` or
    `journalctl --since noon`) turn into a simple ordered scan in RocksDB.

The object headers for everything but entries are synthesized at
read-time; in particular, the linked-list that would be represented by
next_field_offset is just filled-in with a TOKEN_FIELD_NEXT sentinel
value.  This bakes in the assumption that libsystemd's use of these is
always a sequential scan, which seems to be the case today ... but may
be incorrect.

With ZSTD compression enabled in RocksDB, the initial results are
promising; with 130MB of test data (as measured by the `export` output
mode), this results in 89MB of a RocksDB file; after a RocksDB
compaction cycle, this is further compressed to 34MB.  In the
traditional binary journal-file format, this data would have taken
approximately 120MB on disk.

NOTE: This does not currently pass all the tests; rotation rename()s the
file out from under RocksDB, and since RocksDB has a userspace list of
open files, it thinks `test.journal` (no timestamp) is already opened
and refuses to re-open it.